ROS implementation 	modified:   libopendavinci/CMakeLists.txt 	modified:   libopendavinci/include/core/base/ROSDeserializer.h 	modified:   libopendavinci/include/core/base/ROSSerializer.h 	modified:   libopendavinci/src/core/base/ROSDeserializer.cpp 	modified:   libopendavinci/src/core/base/ROSSerializer.cpp 	modified:   libopendavinci/src/core/base/SerializationFactory.cpp 	modified:   libopendavinci/src/core/data/Container.cpp 	modified:   libopendavinci/src/core/data/control/VehicleControl.cpp

+++ OpenDaVINCI-msv/libopendavinci/include/core/base/ROSSerializer.h
@@ -77,7 +77,7 @@ namespace core{
                             
                 ostream &m_out;
                 stringstream m_buffer;
-                                uint8_t m_size;
+                                uint32_t m_size;
         };
     }
 } // core::base
